02-use-cases/01-conversational-agents/A2A-multi-agent-incident-response/web_search_openai_agents/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
a2a-sdk==0.3.8
    # via web-search-openai-agents (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
autopep8==2.3.2
    # via bedrock-agentcore-starter-toolkit
bedrock-agentcore==0.1.7
    # via
    #   web-search-openai-agents (pyproject.toml)
    #   bedrock-agentcore-starter-toolkit
bedrock-agentcore-starter-toolkit==0.1.24
    # via web-search-openai-agents (pyproject.toml)
boto3==1.40.50
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
botocore==1.40.50
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   boto3
    #   s3transfer
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via prance
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via griffe
cryptography==46.0.5
    # via pyjwt
distro==1.9.0
    # via openai
docstring-parser==0.17.0
    # via bedrock-agentcore-starter-toolkit
google-api-core==2.26.0
    # via a2a-sdk
google-auth==2.41.1
    # via google-api-core
googleapis-common-protos==1.70.0
    # via google-api-core
griffe==1.14.0
    # via openai-agents
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   a2a-sdk
    #   bedrock-agentcore-starter-toolkit
    #   mcp
    #   openai
    #   tavily-python
httpx-sse==0.4.3
    # via
    #   a2a-sdk
    #   mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via bedrock-agentcore-starter-toolkit
jiter==0.11.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via
    #   mcp
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via openapi-spec-validator
jsonschema-specifications==2025.9.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.12.0
    # via openapi-spec-validator
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mcp==1.23.0
    # via openai-agents
mdurl==0.1.2
    # via markdown-it-py
openai==1.109.1
    # via openai-agents
openai-agents==0.3.3
    # via web-search-openai-agents (pyproject.toml)
openapi-schema-validator==0.6.3
    # via openapi-spec-validator
openapi-spec-validator==0.7.2
    # via bedrock-agentcore-starter-toolkit
openinference-instrumentation==0.1.40
    # via openinference-instrumentation-openai-agents
openinference-instrumentation-openai-agents==1.3.0
    # via web-search-openai-agents (pyproject.toml)
openinference-semantic-conventions==0.1.24
    # via
    #   openinference-instrumentation
    #   openinference-instrumentation-openai-agents
opentelemetry-api==1.38.0
    # via
    #   openinference-instrumentation
    #   openinference-instrumentation-openai-agents
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.59b0
    # via openinference-instrumentation-openai-agents
opentelemetry-sdk==1.38.0
    # via openinference-instrumentation
opentelemetry-semantic-conventions==0.59b0
    # via
    #   openinference-instrumentation-openai-agents
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   prance
pathable==0.4.4
    # via jsonschema-path
prance==25.4.8.0
    # via bedrock-agentcore-starter-toolkit
prompt-toolkit==3.0.52
    # via
    #   bedrock-agentcore-starter-toolkit
    #   questionary
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.1
    # via
    #   a2a-sdk
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
py-openapi-schema-to-json-schema==0.0.3
    # via bedrock-agentcore-starter-toolkit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.14.0
    # via autopep8
pycparser==3.0
    # via cffi
pydantic==2.12.0
    # via
    #   a2a-sdk
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.41.1
    # via pydantic
pydantic-settings==2.11.0
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.11.0
    # via mcp
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.3
    # via
    #   bedrock-agentcore-starter-toolkit
    #   jsonschema-path
questionary==2.1.1
    # via bedrock-agentcore-starter-toolkit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
regex==2025.9.18
    # via tiktoken
requests==2.32.5
    # via
    #   bedrock-agentcore-starter-toolkit
    #   google-api-core
    #   jsonschema-path
    #   openai-agents
    #   prance
    #   tavily-python
    #   tiktoken
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.2.0
    # via
    #   bedrock-agentcore-starter-toolkit
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruamel-yaml==0.18.15
    # via
    #   bedrock-agentcore-starter-toolkit
    #   prance
ruamel-yaml-clib==0.2.14
    # via ruamel-yaml
s3transfer==0.14.0
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==3.0.2
    # via mcp
starlette==0.49.1
    # via
    #   bedrock-agentcore
    #   mcp
tavily-python==0.7.12
    # via web-search-openai-agents (pyproject.toml)
tiktoken==0.12.0
    # via tavily-python
toml==0.10.2
    # via bedrock-agentcore-starter-toolkit
tqdm==4.67.1
    # via openai
typer==0.19.2
    # via bedrock-agentcore-starter-toolkit
types-requests==2.32.4.20250913
    # via openai-agents
typing-extensions==4.15.0
    # via
    #   anyio
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   mcp
    #   openai
    #   openai-agents
    #   openinference-instrumentation-openai-agents
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.0
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   botocore
    #   requests
    #   types-requests
uvicorn==0.37.0
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   mcp
wcwidth==0.2.14
    # via prompt-toolkit
wrapt==1.17.3
    # via
    #   openinference-instrumentation
    #   openinference-instrumentation-openai-agents
    #   opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
